Implementing REMP segments users API call.
The call depends on a fact, that there's always a fake/dummy event
for every user in every measurement. The reason for this is that one
cannot list users which are not in the database yet.

This responsibility is currently planned for SSO, which will store
registration and login events anyway.

remp/remp#28