Browser's cache should be incremented only when all flags are valid.
Flags are setting on a segment rule to determine whether tracked event
can be used or not. Right now we only support "_article" flag. If not set,
all pageview events are used. If set to "1", only article pageviews are used.
If set to "0", only non-article pageviews are used. The flag is tracked
internally by tracker and values are known to the system.

Segment cache incremented in browser was being incremented always, no
matter the setting of the flags. This commit adds validation of flags
and passing of flags from Segments API to the remplib JS.

remp/remp#117