Refactoring relation between campaigns and banners, marking banners of active campaigns
The original implementation of Banner::campaigns() used simple "orWhere"
condition which worked fine until eager loading as attempted. It started
to generate wrong queries and preloading campaigns incorrectly.

Because of that I've changed the database and model to use pivot table
and alter the mapping so no frontend changes needed to be made.

remp/remp#158