Removing mail logs from template detail, showing template preview in detail
Log was slowing down detail load by seconds (due to SELECT COUNT from
mail_logs) and HTML preview will give user better understanding of which
email was just opened.

Log was only commented so it can be returned once the mail_logs table
is solved in a way that cardinality won't infer with performance that much.