Bugfixing broken A/B test in MailWorker, improving performance of duplicite check
The A/B test got broken after batch sending introduction. The template
was not properly selected for each email as generated by mail:process-job,
but for the first email within the batch. Now for each template the batch
of emails is split to smaller batches with the same mail templates.

Duplicate sending is now being checked for the whole batch of emails,
not by single email due to lower performance of possibly high-cardinality
mail_logs table.