Update schedule with campaign activate-toggle
New toggle was added to campaign list (remp/remp#150) to quickly activate
or deactivate campaign. Now it also will take care of schedules.

What happens when toggle is used.

If campaign is not active, activate it:
 - change campaign's `active` flag to true,
 - create new schedule with status executed (it wasn't planned).

If campaign is active, deactivate it:
 - change campaign's `active` flag to false,
 - stop all running or planned schedules.

Not using existing status pause to pause and unpause existing schedules.
Reason for that is too many ways how to do it. (Eg. Pause all schedules
and then unpause all? Or reactivate only last one? What if none of them
has now end time in future?).

remp/remp#179