Removing throttling from SSO
The reason for removal is that each service is expected to verify
JWT token on each of its calls. Therefore there's an expectation,
that most requests will come anonymously from a single IP.

These requests would have been incorrectly throttled. As a simple
solution we remove the throttling for now.