Using direct airbrake notifier within exception handler due to phpbrake bug
When PhpBrake tries to log the exception, it ignores the stack trace
of exception and builds new trace. Unfortunately when Monolog logger is
used, this trace is pointing to the logger and not to the point where
the error occurred.

Until the bug is fixed, we notify the airbrake directly instead of using
Monolog logging. File logging was kept.

https://github.com/airbrake/phpbrake/issues/87