Linking tracy's showBar option to APP_DEBUG env variable
When Tracy is enabled and the bar is not shown, tracy.bar route is not
defined but used by the library. Due to that the error handling fails
and exceptions are not properly processed.

The bug report was issued to the owner of the library. Once it's fixed,
we can revert the configuration back so we can have split debugging
and Tracy bar showing.

https://github.com/recca0120/laravel-tracy/issues/48