Alias mapping in segments API reads mappings from all indices
The original implementation read mapping only from the latest index
(sorted alphabetically). This occasionally caused issues as not all
fields were present in the latest index and mapping is not enforced
when events are being tracked.

As these fields still might have been selected by caller, map caching
is now looking at older alias indices as well and if it finds field
that was not mapped yet and adds it to the mapping.

remp/novydenik#402