Adding caching of requests for Beam concurrents call
The first instance of caching caches response for the simplest Beam call.
It rounds down the requested timeframe to 10 seconds so it can reuse
the same cached value for the requests.

Value is cached for 60 seconds right now, but after 10 seconds it shouldn't
be accessed as new cache key is generated and used by the clients.

remp/remp#416