Using banner variant UUID as a key to iterate and display stats
Before we used regular ID and combined it with UUID found in Elastic
data. This didn't work correctly if publisher mistakenly tracked
different UUID than hours and we attempted to load non-existing
variant based on the wrong UUID.

Since we want to display stuff based on the UUID, it should be the
primary key to iterate and search with.