Using stream JSON decoding for CRM segment integration
The memory usage of standad json_decode got very high
when it tried to parse ~10MB of JSON in response body.
As the time is not the issue right now, we'd rather use
implementation using less memory as 256M limit stopped
being enough on the production environment.

This commit switches standard json_decode to halaxa/json-machine
which parses JSON as it reads and doesn't keep full structure
in the memory.

Comparison of ~10MB of decoding response with list of users:

1) json_decode: 120MB of memory used, 12 seconds to decode
2) halaxa/json-machine: 9MB of memory used, 80 seconds to decode

This is currently tradeoff we can live with.