Major update of Go dependencies, preparing for Elastic update
* Change in Elastic mapping caching was necessary as Elastic 7.*
is dropping support for type specificiation and _doc is used implicitly
and by default. Implementation was altered in a way that both 6.* and
7.* will still be supported. This is also handling the deprecation
warning which was present since 6.8 and 7.*.

See https://www.elastic.co/guide/en/elasticsearch/reference/master/removal-of-types.html

* Updateable Go packages were updated. This doesn't affect Goa which
remained on the version 1.2.1 and will require more time to update
to the latest version.

  * Update caused extra "AllowNativePasswords" parameter to be configured
  within MySQL configuration in both Segments and Tracker.

  * Update also caused an extra override of "gopkg.in/fsnotify.v1"
  in Gopkg.toml which fixed panic within "dep ensure" command.
  See https://github.com/golang/dep/issues/1799#issuecomment-494835176.

* Index creation in Elastic docker image now creates all indices
(some of them were omitted before)