Removing remplib import, counting on global instance
This was interfering with attempts to use global instance of remplib with all
the configuration present.

Using global remplib is not the best solution, but proper refactoring
and passing dependencies isn't part of the implementation yet.

This commit is only to bugfix existing issue which caused banner closes
and clicks not being tracked properly since the release.