Removing dummy "center" position from Campaign
The position was added there in the past during work on overlay
rectangle banner. It was never defined within position map and
actually used intentionally. Banners positioned with "center"
value were actually within their default position.

As position might be NULL in the database, this commit alters the
request validation and allows position not to be provided.