Custom JS was not being triggered
Custom JS of the Banner was not being triggered. 

Scenario:
1. There was a custom js (you can test with an alert);
2. There was no js include;

In this scenario, the variable this.jsIncludes, for some reason, represents an array of one position with null value. So the condition on line 276 results true, executing the continue statement. As there was only one null position on array, the code above line 281 was never executed because the for didn't have any more elements to execute.

The bug can be easily reproduced adding an alert inside the custom js of a banner. 

I corrected the bug and simplified the algorithm. Hope you enjoy the art.