Bug Correction
Problem:
At release 0.9.1, an Internal Server Error occurs during banner rendering. All banners stop working. The log points to CampaignController, line 549, where the code call count function passing the campaignIds. An error is thrown saying that campaignIds is not an array. In Redis CLI, I executed GET active_campaign_ids and it returned an object and not an array as expected.

Investigation:
After some investigation, I discovered that in the table Schedules from database Campaign, there was two schedules for the same campaign. Inside Campaign.php, method refreshActiveCampaignsCache, the code get all the active or scheduled campaigns in line 249 and after this calls ->unique(), that removes repeated elements in the resulting array. When it happens, an array like [0 => 1, 1 => 2, 2 => 3] (key  => value) turns into [0 => 1, 2 => 3]. The array is not continuous anymore. So when we call Json::encode, it convert the value into an object.

Solution: 
Call array_values($campaignIds) to guarantee that the array keys is continuous. The value now is converted into an array like it is expected.