User/Browser needs to match all segment rules to see the campaign
Refactored logic changed the implementation and started to display
banner when user matched at least one of the segment rules. This
would be a breaking change in contradiction how the segments are
intended to work. Now all the segment rules have to match to see
the banner.

SegmentAggregator is also passed as a constructor dependency.