Adding configuration option to force HTTPS in generated URLs
If application is running on HTTPS publicly, but internal
proxy forwarding convert the request to HTTP, all generated
URLs would contain HTTP instead of HTTPS.

As application wants to generated URLs with HTTPS protocol
to maintain public protocol, we needed to add configuration
option to enfore this behavior.

`FORCE_HTTPS` was added with default value `false`. Same
configuration was already present in Mailer (Nette), but
was missing in other Laravel applications.

remp/remp#619