Attribute conversion source to the medium
There are two approaches to attribute the source:

- SESSION_FIRST finds the session preceeding the conversion
and determines the source based on the source of the first
pageview of that session.
- SESSION_LAST works similarly, but uses the last pageview
of the matched session.

Attribution is done on the fly when conversion is
upserted to the system.

There's also conversions:process-sources command to run
the processing for all unprocesssed conversions.

remp/remp#464