Fix empty (undefined) cookies and storage entries
In Firefox (not tested in another browsers), the method parseUriParams
was creating empty entries in Cookies / Local Storage when
there's not query params.

This was the scenario to reproduce the problem:

1. There's no uri param;
2. Function substring(1) of window.location.search returned
empty "" (line 656);
3. Function query.split('&') was returning an array containg
one "" empty value (line 660);
4. The for iterator (line 664) considered empty "" as a value
and remplib.setStorage(key, item) is called with remplib.setStorage("", undefined);

A conditional checking if variable query is empty or undefined that returns in case of true.

remp2020/remp#81

Signed-off-by: Peter Dulacka <dulacka@gmail.com>