Fix empty Elastic index initialization during installation
After update to v7 it's necessary to create new indexes with
`POST` request instead of `PUT`. Current version silently
failed and proceeded with no index created.

This was not a blocker as index gets created when first event
is hit, however it could have caused some issues if a client
attempted to read events from index that doesn't exist yet.