Handle edge case with empty string as click tracking param
Frontend receives null and interpolates it as empty string
select value. When this comes back to the server, we need
to convert it back to null, otherwise boolean value of ""
(false) is used to store click tracking configuration.

This caused accidental disabling of click tracking on emails
where default behavior was supposed to be used.