Fix Beam dependency issues after update
- Snowplow library is now referencing wrong repository. We
needed to manually point that to the correct one as custom
package. While being there, RefererParser now uses more recent
list of sources from the Snowplow's main repo.
- Disabled default empty value for PAGEVIEW_GRAPH_DATA_SOURCE
environment variable. Having empty value caused charts not to
work as the default didn't kick in - it works only when
the variable is not set at all.