Fix `search.max_buckets` issue on pageviews count aggregation
If the number of aggregated buckets was greater than 10000,
Elastic returned an error. The solution was either to raise
the limit (not recommended) or switch to the composite
aggregations.

With pageview-related aggregations it was possible and
result is now paged per 10K buckets.

Commerce and event aggregations were not migrated yet
since composite aggregation does not support `MinDocCount`
and `ExtendedBounds` to allow empty buckets within the
whole time range.

Other minor changes:

- Fix Interval deprecation warning by swapping it with
FixedInterval implementation.
- Use filter_var to handle boolean variables in PHP
processing of Segments API responses.
- Count histogram items were originally float. This was
unnecessary because "count" can always be only integer.
All related implementations were changed.
- Due to the limitations of composite aggregations, it's
not possible to use offset on histograms (yet). Since we
never used it, we remove it from the APIs for now.

remp/remp#622