Add newsletter rectangle banner template
The banner template is meant to ease scenario
of subscribing to newsletter directly from within
the banner.

If Campaign instance is using Mailer, list of
available newsletters is provided when configuring
the new banner.

The implementation requires proxy to handle the
requests to make sure that backend correctly
handles user registration within CRM. The way
of communication with proxy is highly configurable
and described within README and .env.example file.