Mailer batch processing memory optimizations
The original implementation needed to store all
the batch data within `$userJobOptions`.
This was necessary for doing unreadArticlesResolver
requests in chunks. Unfortunately this array grew
very fast and could hit memory limits.

Because of this, commit adds new one-off implementation
of parameter resolving that will be executed for
each recipient of the batch separately. It might
slow down sending personalized emails a bit, but
it solves memory issues for any kind of batch/segment
size in the future.

remp/helpdesk#116