Change banner template columns with text to text type
Originally there was string (varchar) type which could cause
an exception if a too long text was saved into the banner.

Since we allow HTML/JS/images within the banner text, it's
safe to say 191 characters are not enough for text to be
safely stored and this change of type was necessary.

remp/remp#819