Make sure remp:showtimeReady is emitted at right time
Banners are rendered asynchronously and event was still
emitted before the banners were displayed and before
banner's JS was executed.

For this, we need to pass promise resolve to the banner
JS handler which gets executed at the end of processing.

Custom JS within the banner was executed again asynchronously.
This wasn't really necessary (whole banner is already async)
so we removed it for now for the sake of emitting event
correctly.

remp/web#1393