Fixed slow initialization of grid filters
On some instances and browsers, initialization of select
filters could take more than a minute. It didn't seem to
be 100% replicable, but when it happened in the browser,
it happened all the time.

This commit removes jQuery from the point where we dynamically
create select/options tag which seems to help for the affected
browsers.