Fix missing data in pageviews aggregation
Segments API was building request to elastic which ignored
aggregation of buckets which didn't hold values for all
groupped fields.

We had fallback in place for string values which helped
with aggregation of records with missing string value,
but we didn't have one for boolean values.

If we aggregated data (e.g. pageviews) and the boolean value
was missing (e.g. "subscriber" flag), Elastic wouldn't
return the aggregated count for these records.

Bool values now default to "false" which helps Elastic
to trigger the aggregation.

remp/remp#869