Fix unresolved `settings` and `unsubscribe` template variables
They were completely missing if system was configured to use Mailgun
mailer with batch sending.

Please be aware that there's still known issue that the links do not
receive tracking (RTM) params if batch sending is used. This is
because Mailer only replaces URLs within anchor tags. When using
Mailgun with batch sending, replaced doesn't see URL within the anchor
tag, but this:

  <a href="%recipient.unsubscribe%"></a>

Such use is ignored, because we rely on valid URL within the tag.

remp/remp#879