Fixed possibly invalid aggregation of conversion data
The issue was caused by invalid storing of time columns in the
aggregated data. The times were not forced to the UTC and application
stored them in their local timezone (stating it's the UTC).

Because of this, all of the time-columns in the aggregated data
(events prior to the conversion) might have been off couple of hours.

Commit fixes the issue and truncates all of the invalid data.
It also sets the flags of conversions to the state so they're
processed again. The fix forces BaseModel class for all of the models
so we can enforce UTC timezone before storing the date to the DB.

remp/remp#464