Fixed issue with API calls generating PHP sessions, bump dependencies
Session was generated on each request, causing session store to be
overloaded with records. The session was created by a tracy (debug)
library on every request, even though Laravel doesn't touch session
on API requests.

  https://github.com/recca0120/laravel-tracy/blob/3ea29da4224c6700175552c5df482e90ca4f4fa3/src/DebuggerManager.php#L189

This library was unnecessarily included in the "require" block of
the composer.json and was moved to "require-dev".

remp/remp#954