Fix issues with control group tracking and data persistence
If user was supposed to receive control group variant
(without banner), there would be a JS error that prevented
data to be tracked correctly.

Also the variant wouldn't be stored in the storage data
and the next pageview would have to determine the variant
again (instead of reusing the one already selected
for the user).

remp/remp#980