Fix ignoring of segment's `active` flag in segments DB
Cache method and listing worked correctly only with active
segments. However if the direct segment access didn't find
the segment in the cached map, it tried to fetch it from the DB.
The fetch was missing active=1 condition which caused
inactive segments to appear in the listings and in the calls
targeted only for active segments.

remp/remp#1007