Add option to configure step of pageviews:aggregate-articles-views command
This is toavoid Elasticsearch's "Trying to create too many buckets" error,
which can happen to installations with more data.

Elasticsearch needs to create a bucket for each user/browser/article combination. If the number of buckets is greater than allowed
(defaults to 25K), it ends up with error.

remp/remp#1050