Explicitly casting user ids to string
This is to improve query performance. We used user_ids as array keys
in the implementations and PHP converted them to integers.
When integer is passed as a query parameter to the string column,
index is not used and query performance is hit drastically.

We needed to explicitly cast these external IDs as strings
to the index is used correctly.

remp/remp#1088