Fixed fallback for non-batch sending in mail:worker command
If the worker was run with `--batch` flag and the configured
mailer didn't support batch sending, the fallback would execute
incorrect branch and the sending would end up with error.

This is because the $email would already have multiple configured
recipients and the Sender::send() method doesn't allow that.

Fix ensures that there's only one recipient in the $email when
the selected Mailer doesn't support batch sending.

remp/remp#1139