Fix conversion attribution issue
The bug could possibly causing conversions happening early
after the job was created. It happened because the time of
conversion could be in the different timezone as the application.

Nette doesn't change the timezone automatically, so in order
to make sure the dates are queried/inserted in the correct
timezone, they're now being processed by the DateFieldsProcessorTrait.

remp/remp#1144