Fix issue with worker:mail getting stuck
There is an unreproduced edge case where queue can get stuck
in "no longer top priority, switching" message. This is caused
by inconsistency between MySQL (batch is in one of the "processing"
statuses) and Redis (batch is already in the priority queues).

If this happens again, let mail worker to clean this mess up
and remove the queue from Redis cache.

remp/remp#1189