Fix issues with Mailer batch processing and Redis queues
1) BatchEmailGenerator at the end updated the priority-mail-queues
record in Redis. This happened even when the batch wasn't scheduled
to be sent right away (via "ready_to_process" state).

MailWorker expected priority-mail-queues to be batches ready for
handling. This fix covers scenario:

- Batch set to "ready_to_process"
- Generator populated batch to the mail cache (Redis)
- Generator updated priority-mail-queues
- Mail worker realized, that newly created batch can be a top priority
batch and switched to that. Right after that it realized it's not being
sent (it never was) and cleared things up.

2) When the job processing fails, Mailer changes the status from
"processing" to one of the "ready_to_process" statuses.

The failed job processing could already write something to the cache
and this needs to be cleared before we try to process the batch again.
Otherwise duplicate records would be present in the cache and
the mail workers would need to handle this batch longer than necessary.

remp/remp#1195