Disable subscribe tracking in Mailer module
The tracking is dependent on the Beam Tracker configuration
and that is not present by default. Clean installation of
Mailer would crash because of the missing dependency.

Enabling the tracker is now up to each Mailer installation.

The hint to enable them is now moved to the Mailer Skeleton
repository - config.local.neon.example.