Fix build process of Go applications
We use older Goa v1 to define and generate Go APIs. Goa v1
has a dependency for github.com/armon/go-metrics. The library
was renamed recently and go modules used to resolve
the dependencies now won't allow installation of the library,
because Goa requests "armon" library, but Go finds "hashicorp"
library.

We ended up forking the library to github.com/rootpd/goa and
adding "replace" statement to go.mod file which handles the
transition.

https://github.com/rootpd/goa/blob/v1/go.mod#L5

Unfortunately once go.mod contains replace statement, go modules
won't allow using go install to install the package. Because of
that, we needed to replace go install with more raw git checkout
and local build of goagen.

Once we are on the latest version of Goa, we should revert the
process back to the go install.

remp/remp#1275