Fix Mailer configs to not require CRM and Beam with first install
Few hermes handlers registered in `config.neon` require CRM and Beam
client to be initialized. Which shouldn't be required when Mailer is
installed for the first time.

Services are now only in `config.local.neon.example` commented out
with `crmClient` and `Tracker\Beam` ready to be uncommented and used
(env variables have to be properly set).