Align Mailer hermes queues priority
The default Hermes queue for asynchronous events is
now "medium priority queue" instead of "low priority
queue".

Use of the low priority as the default could cause
starvation of some events emitted with the default
priority, because it could get mixed with the higher
amount of low-priority events.

The medium priority queue as the default is the
reasonable option.

remp/remp#1342