URL parser generator's segment is now optional
If not provided, Mailer's segment with subscribers
of selected mail type is used as a default.

This change should speed up the generation as it's
not necessary to use all_users/everyone segments
anymore. Use of segment in URL parser generator
should be limited to selection of subset of list
subscribers.

remp/crm#2973