Enforce empty type to be used for Elasticsearch mapping
This attempts to fix compatibility issues between v7 and v8.
If no type is used, v7 library appends "/_all" to the request
which is not supported by v8. By using empty type, the request
should be compatible with both of the versions.

Target request is /pageviews/_mapping/, original request
was pointing to /pageviews/_mapping/_all.

Elasticsearch 8 was ending up with this error:

{"error":"no handler found for uri [/pageviews/_mapping/_all] and method [GET]"}

remp/remp#1365