Update the init script for Elasticsearch Docker image
The script now creates indices via Index Lifecycle
Management (ILM) with some default configuration
settings, which mirror the defaults we used to have
in the commands for index managements:

- Rollovers after 4G of data.
- Data retention based on the index, expects the
data to be exported before the retention hits.

This change ensures that Elasticsearch itself
executes the rollovers and data retention policies,
so we don't have to anymore.

remp/remp#1419