Fix routing conflict between web and API routes
The web and API routes had the same alias, due to
the changes made in routing in Laravel 11. The latest
Laravel 12 changed something and the API route was
resolved as first and won. It caused web links to point
to API instead of web controllers.

This adds a prefix to all API routes so they're not
in conflict with web routes anymore.

remp/remp#1447

(cherry picked from commit 0c07c73e8e8339aef16969d6623291c9bfb4980c)