Remove FilterLoader in favor of direct filter registration
Latte allows registration of filters directly and the
use of FilterLoader is deprecated since Latte 3.1.
Since the deprecation is more or less strict, we're
dropping support for FilterLoader in favor of direct
registration.

Any latte filters should now be registered in DI through
nette.latteFactory service by using addFilter() method.
The signature is the same as signature of FilterLoader::register
used previously.